ensure the cause is an error
