+++ packages/server/src/error/TRPCError.ts
@@ -1,5 +1,5 @@
 import { TRPC_ERROR_CODE_KEY } from '../rpc/codes';
-import { isObject } from './utils';
+import { getCauseFromUnknown } from '../shared/getCauseFromUnknown';
 
 export function getTRPCErrorFromUnknown(cause: unknown): TRPCError {
   if (cause instanceof TRPCError) {
@@ -21,37 +21,6 @@ export function getTRPCErrorFromUnknown(cause: unknown): TRPCError {
   }
 
   return trpcError;
-}
-
-class UnknownCauseError extends Error {
-  [key: string]: unknown;
-}
-
-function getCauseFromUnknown(cause: unknown): Error | undefined {
-  if (cause instanceof Error) {
-    return cause;
-  }
-
-  const type = typeof cause;
-  if (type === 'undefined' || type === 'function' || cause === null) {
-    return undefined;
-  }
-
-  // Primitive types just get wrapped in an error
-  if (type !== 'object') {
-    return new Error(String(cause));
-  }
-
-  // If it's an object, we'll create a synthetic error
-  if (isObject(cause)) {
-    const err = new UnknownCauseError();
-    for (const key in cause) {
-      err[key] = cause[key];
-    }
-    return err;
-  }
-
-  return undefined;
 }
 
 export class TRPCError extends Error {
